+++ src/KardFinancial.Test/Unit/MockServer/Organizations/ContentStrategies/GetTest.cs
@@ -30,9 +30,7 @@ public class GetTest : BaseMockServerTest
                 "merchantExclusions": [
                   "merchantExclusions",
                   "merchantExclusions"
-                ],
-                "createdAt": "2024-01-15T09:30:00.000Z",
-                "lastModified": "2024-01-15T09:30:00.000Z"
+                ]
               }
             }
             """;