+++ src/KardFinancial.Test/Unit/MockServer/Organizations/ContentStrategies/UpdateTest.cs
@@ -55,9 +55,7 @@ public class UpdateTest : BaseMockServerTest
                 "merchantExclusions": [
                   "merchantExclusions",
                   "merchantExclusions"
-                ],
-                "createdAt": "2024-01-15T09:30:00.000Z",
-                "lastModified": "2024-01-15T09:30:00.000Z"
+                ]
               }
             }
             """;