+++ src/KardFinancial.Test/Unit/MockServer/Organizations/Placements/ListTest.cs
@@ -22,9 +22,7 @@ public class ListTest : BaseMockServerTest
                     "name": "name",
                     "organizationId": "organizationId",
                     "availableSlots": 1,
-                    "contentStrategyId": "contentStrategyId",
-                    "createdAt": "2024-01-15T09:30:00.000Z",
-                    "lastModified": "2024-01-15T09:30:00.000Z"
+                    "contentStrategyId": "contentStrategyId"
                   }
                 },
                 {
@@ -34,9 +32,51 @@ public class ListTest : BaseMockServerTest
                     "name": "name",
                     "organizationId": "organizationId",
                     "availableSlots": 1,
-                    "contentStrategyId": "contentStrategyId",
-                    "createdAt": "2024-01-15T09:30:00.000Z",
-                    "lastModified": "2024-01-15T09:30:00.000Z"
+                    "contentStrategyId": "contentStrategyId"
+                  }
+                }
+              ],
+              "included": [
+                {
+                  "type": "contentStrategy",
+                  "id": "id",
+                  "attributes": {
+                    "name": "name",
+                    "organizationId": "organizationId",
+                    "filter": "NEWLY_LIVE",
+                    "categories": [
+                      "Arts & Entertainment",
+                      "Arts & Entertainment"
+                    ],
+                    "categoryExclusions": [
+                      "Arts & Entertainment",
+                      "Arts & Entertainment"
+                    ],
+                    "merchantExclusions": [
+                      "merchantExclusions",
+                      "merchantExclusions"
+                    ]
+                  }
+                },
+                {
+                  "type": "contentStrategy",
+                  "id": "id",
+                  "attributes": {
+                    "name": "name",
+                    "organizationId": "organizationId",
+                    "filter": "NEWLY_LIVE",
+                    "categories": [
+                      "Arts & Entertainment",
+                      "Arts & Entertainment"
+                    ],
+                    "categoryExclusions": [
+                      "Arts & Entertainment",
+                      "Arts & Entertainment"
+                    ],
+                    "merchantExclusions": [
+                      "merchantExclusions",
+                      "merchantExclusions"
+                    ]
                   }
                 }
               ],